move LinkableSymbol identifer/analyzer to their own file
